+++ test/unit/owlImport.characteristics.test.ts
@@ -18,7 +18,7 @@ import { importCharacteristics } from '../../src/modules/ontology/importDispatch
 import { SchemaRegistry } from '../../src/modules/registry/SchemaRegistry.js';
 import { Curie } from '../../src/modules/rdf/Curie.js';
 import { Terms } from '../../src/modules/rdf/Terms.js';
-import { DEFAULT_PREFIXES } from '../../src/constants/PREFIXES.js';
+import { STANDARD_PREFIXES } from '../../src/constants/STANDARD_PREFIXES.js';
 import type { QuadInterface } from '../../src/interfaces/Quad.js';
 import type { OwlImportContext } from '../../src/interfaces/OwlImport.js';
 import { SchemaGraph } from '../../src/modules/graph/SchemaGraph.js';
@@ -54,7 +54,7 @@ const OWL_IRREFLEXIVE_CURIE = 'owl:IrreflexiveProperty';
 // ---------------------------------------------------------------------------
 
 function makeCtx(allPropertyIris?: ReadonlySet<string>, quads: QuadInterface[] = []): OwlImportContext {
-  const curie = new Curie(DEFAULT_PREFIXES);
+  const curie = new Curie(STANDARD_PREFIXES);
   const unsupported: Array<{
     'axiomIri': string;
     'subjectIri': null | string;
@@ -70,7 +70,7 @@ function makeCtx(allPropertyIris?: ReadonlySet<string>, quads: QuadInterface[] =
     'isDatatype': () => {
       return false;
     },
-    'prefixes': DEFAULT_PREFIXES,
+    'prefixes': STANDARD_PREFIXES,
     'reportUnsupported': (axiomIri, subjectIri) => {
       unsupported.push({
         axiomIri,